PERA-4068 Delegate swap percentage calculation to backend by using new endpoint (#529)

+++ app/src/main/kotlin/com/algorand/android/ui/swap/widget/viewmodel/SwapWidgetViewModel.kt
@@ -28,7 +28,7 @@ interface SwapWidgetViewModel : StateViewModel<ViewState> {
 
     fun getAmountInputFlow(): StateFlow<SwapAmountInput.Input>
 
-    fun setAmountByPercentage(swapDetails: SwapViewModel.SwapDetails, percentage: Int)
+    fun setAmountByPercentage(swapDetails: SwapViewModel.SwapDetails, percentage: Float)
 
     fun initWidget(
         swapDetailsFlow: Flow<SwapViewModel.SwapDetails>,
